Add dockerfile, docker-compose.yml.example and version endpoint
