+++ docker-compose.yml.example
@@ -1,0 +1,34 @@
+services:
+  backend:
+    build:
+      context: .
+      dockerfile: dockerfile
+    ports:
+      - '3000:3000'
+    volumes:
+      - .:/app
+    environment:
+      - NODE_ENV=development
+    restart: always
+    networks:
+      - pdesa
+
+  frontend:
+    build:
+      context: ../pdesa-frontend
+      dockerfile: dockerfile
+      args:
+        - VITE_SELF_VERSION=hola
+        - VITE_BACKEND_URL=http://localhost:3000/
+    ports:
+      - '80:80'
+    volumes:
+      - .:/app
+    environment:
+      - VITE_BACKEND_URL=http://localhost:3000/
+    restart: always
+    networks:
+      - pdesa
+
+networks:
+  pdesa:
