+++ src/app/app.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { ConfigModule } from '@nestjs/config';
-import configuration from './config/configuration';
+import configuration from '../config/configuration';
 
 @Module({
   imports: [ConfigModule.forRoot({ isGlobal: true, load: [configuration] })],
