+++ src/app/response/version.response.ts
@@ -1,0 +1,10 @@
+import { ApiProperty } from '@nestjs/swagger';
+
+export class VersionResponse {
+  @ApiProperty()
+  version: string;
+
+  constructor(version: string) {
+    this.version = version;
+  }
+}
