+++ src/config/configuration.ts
@@ -5,6 +5,7 @@ dotenv.config({});
 const configuration = {
   app: {
     port: process.env.PORT ? parseInt(process.env.PORT, 10) : 3000,
+    version: process.env.SELF_VERSION ?? '-',
   },
 };
