+++ test/e2e/app.e2e-spec.ts
@@ -2,7 +2,7 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { INestApplication } from '@nestjs/common';
 import request from 'supertest';
 import { App } from 'supertest/types';
-import { AppModule } from '../../src/app.module';
+import { AppModule } from '../../src/app/app.module';
 
 describe('AppController (e2e)', () => {
   let app: INestApplication<App>;
